src/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "deepeval>=3.9.3",
    "dspy>=3.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
